refactor: Header.js => ts 로 migration
