+++ src/App.js
@@ -1,4 +1,4 @@
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 import MovieList from "./components/MovieList";
 
 class App {
